add a lot of thing

+++ src/pages/Success/Success.js
@@ -1,5 +1,6 @@
 import axios from "axios";
 import {
+  addDoc,
   collection,
   deleteDoc,
   doc,
@@ -22,10 +23,10 @@ export default function Success() {
   const location = useLocation();
   // const { documents: transactions} = useCollection()
   const currentPath = location.search;
-  console.log(currentPath)
-  const [transaction, setTransaction] = useState();
+  console.log(currentPath);
+  const [transaction, setTransaction] = useState({});
   const { documents: License } = useCollection("user", ["uid", "==", user.uid]);
-
+  
   const navigate = useNavigate();
   useEffect(() => {
     const colRef = collection(db, "transaction");
@@ -39,37 +40,58 @@ export default function Success() {
       setTransaction(order);
     });
   }, []);
-
   useEffect(() => {
+    
     if (transaction) {
       console.log(transaction);
       if (transaction.length > 0) {
         if (currentPath !== "?error=501") {
           if (License.length > 0) {
-            const order = localStorage.getItem("amount");
-            const currentDate = moment().format("MM-DD-YYYY");
-            const amount = parseInt(order);
-            const numberOfMonths = amount / 5000;
-            const newDate = moment(currentDate)
-              .add(numberOfMonths, "months")
-              .format("MM-DD-YYYY");
+            if(transaction[0].description === "Licencja"){
+              console.log(transaction);
+              const order = transaction[0].products[0].unitPrice;
+              const currentDate = moment().format("MM-DD-YYYY");
+              const amount = parseInt(order);
+              const numberOfMonths = amount / 5000;
+              const newDate = moment(currentDate)
+                .add(numberOfMonths, "months")
+                .format("MM-DD-YYYY");
 
-            const docRef = doc(db, "user", License[0].id);
-            setDoc(docRef, {
-              license: "full-license",
-              uid: user.uid,
-              expireDate: newDate,
-            });
-            const ref = doc(db, "transaction", transaction[0].id);
-            deleteDoc(ref);
-            
+              const docRef = doc(db, "user", License[0].id);
+              setDoc(docRef, {
+                license: "full-license",
+                uid: user.uid,
+                expireDate: newDate,
+              });
+              const historyRef = collection(db, "history");
+              addDoc(historyRef, {
+                uid: user.uid,
+                type: transaction[0].description,
+                orderId: transaction[0].orderId,
+                products: transaction[0].products,
+                date: Date.now()
+              })
+              const ref = doc(db, "transaction", transaction[0].id);
+              deleteDoc(ref);
+            }
+            else {
+            const historyRef = collection(db, "history");
+              addDoc(historyRef, {
+                uid: user.uid,
+                type: transaction[0].description,
+                orderId: transaction[0].orderId,
+                products: transaction[0].products,
+                date: Date.now()
+              })
+              const ref = doc(db, "transaction", transaction[0].id);
+              deleteDoc(ref);
+          } 
           }
         }
       } else {
-        if(transaction[0]){
-        const ref = doc(db, "transaction", transaction[0].id);
-        deleteDoc(ref);
-        
+        if (transaction[0]) {
+          const ref = doc(db, "transaction", transaction[0].id);
+          deleteDoc(ref);
         }
       }
     } else {
@@ -78,7 +100,7 @@ export default function Success() {
       }, 1000);
     }
   }, [userDate, transaction]);
-
+  
   setTimeout(() => {
     navigate("/account");
   }, 1500);
